finish testing for aggregator
