+++ service/worker/scanner/executions/workflows_test.go
@@ -277,114 +277,3 @@ func (s *workflowsSuite) TestScannerWorkflow_Failure_ScanShard() {
 	s.True(env.IsWorkflowCompleted())
 	s.Error(env.GetWorkflowError())
 }
-
-func (s *workflowsSuite) TestFlattenShards() {
-	testCases := []struct {
-		input    Shards
-		expected []int
-	}{
-		{
-			input: Shards{
-				List: []int{1, 2, 3},
-			},
-			expected: []int{1, 2, 3},
-		},
-		{
-			input: Shards{
-				Range: &ShardRange{
-					Min: 5,
-					Max: 10,
-				},
-			},
-			expected: []int{5, 6, 7, 8, 9},
-		},
-	}
-	for _, tc := range testCases {
-		s.Equal(tc.expected, flattenShards(tc.input))
-	}
-}
-
-func (s *workflowsSuite) TestShardResultAggregator() {
-	agg := newShardScanResultAggregator([]int{1, 2, 3})
-	expected := &shardScanResultAggregator{
-		reports: map[int]common.ShardScanReport{},
-		status: map[int]ShardStatus{
-			1: ShardStatusRunning,
-			2: ShardStatusRunning,
-			3: ShardStatusRunning,
-		},
-		aggregation: AggregateScanReportResult{
-			CorruptionByType: make(map[common.InvariantType]int64),
-		},
-		corruptionKeys: make(map[int]common.Keys),
-	}
-	s.Equal(expected, agg)
-	report, err := agg.getReport(1)
-	s.Nil(report)
-	s.Equal("shard 1 has not finished yet, check back later for report", err.Error())
-	report, err = agg.getReport(5)
-	s.Nil(report)
-	s.Equal("shard 5 is not included in the shards that will be processed", err.Error())
-	firstReport := common.ShardScanReport{
-		ShardID: 1,
-		Stats: common.ShardScanStats{
-			ExecutionsCount:  10,
-			CorruptedCount:   3,
-			CheckFailedCount: 1,
-			CorruptionByType: map[common.InvariantType]int64{
-				common.HistoryExistsInvariantType:        2,
-				common.OpenCurrentExecutionInvariantType: 1,
-			},
-			CorruptedOpenExecutionCount: 1,
-		},
-		Result: common.ShardScanResult{
-			ShardScanKeys: &common.ShardScanKeys{
-				Corrupt: &common.Keys{
-					UUID: "test_uuid",
-				},
-			},
-		},
-	}
-	agg.addReport(firstReport)
-	expected.status[1] = ShardStatusSuccess
-	expected.reports[1] = firstReport
-	expected.aggregation.ExecutionsCount = 10
-	expected.aggregation.CorruptedCount = 3
-	expected.aggregation.CheckFailedCount = 1
-	expected.aggregation.CorruptionByType = map[common.InvariantType]int64{
-		common.HistoryExistsInvariantType:        2,
-		common.OpenCurrentExecutionInvariantType: 1,
-	}
-	expected.aggregation.CorruptedOpenExecutionCount = 1
-	expected.corruptionKeys = map[int]common.Keys{
-		1: {
-			UUID: "test_uuid",
-		},
-	}
-	s.Equal(expected, agg)
-	agg.addReport(firstReport)
-	s.Equal(expected, agg)
-	report, err = agg.getReport(1)
-	s.NoError(err)
-	s.Equal(firstReport, *report)
-	secondReport := common.ShardScanReport{
-		ShardID: 2,
-		Stats: common.ShardScanStats{
-			ExecutionsCount:  10,
-			CorruptedCount:   3,
-			CheckFailedCount: 1,
-			CorruptionByType: map[common.InvariantType]int64{
-				common.HistoryExistsInvariantType:        2,
-				common.OpenCurrentExecutionInvariantType: 1,
-			},
-			CorruptedOpenExecutionCount: 1,
-		},
-		Result: common.ShardScanResult{
-			ControlFlowFailure: &common.ControlFlowFailure{},
-		},
-	}
-	agg.addReport(secondReport)
-	expected.status[2] = ShardStatusControlFlowFailure
-	expected.reports[2] = secondReport
-	s.Equal(expected, agg)
-}
